gmsol_solana_utils::transaction_builder

Struct TransactionBuilder

pub struct TransactionBuilder<'a, C, T = ()> { /* private fields */ }

A builder for a transaction.

Implementations

impl<'a, C: Deref<Target = impl Signer> + Clone> TransactionBuilder<'a, C>

pub fn new(program_id: Pubkey, cfg: &'a Config<C>) -> Self

Create a new transaction builder.

pub fn merge(self, other: Self) -> Self

Merge other TransactionBuilder. The rpc fields will be empty after merging, i.e., take_rpc will return None.

Panics

Return if there are any errors.

Notes

All options including cluster, commiment and program_id will still be the same of self after merging.

pub fn try_merge(&mut self, other: &mut Self) -> Result<()>

Merge other TransactionBuilder. The rpc fields will be empty after merging, i.e., take_rpc will return None.

Errors

Return error if the payers are not the same.

Notes

When success, the other will become a empty TransactionBuilder.

All options including cluster, commiment, and program_id will still be the same of self after merging.

pub fn into_bundle_with_options( self, options: BundleOptions, ) -> Result<BundleBuilder<'a, C>>

Convert into a BundleBuilder with the given options.

pub fn into_bundle(self) -> Result<BundleBuilder<'a, C>>

Convert into a BundleBuilder.

impl<'a, C: Deref<Target = impl Signer> + Clone, T> TransactionBuilder<'a, C, T>

pub fn payer<C2>(self, payer: C2) -> TransactionBuilder<'a, C2, T>

Set payer.

pub fn get_payer(&self) -> Pubkey

Get the pubkey of the payer.

pub fn cluster(self, url: impl AsRef<str>) -> Result<Self>

Set cluster.

pub fn options(self, options: CommitmentConfig) -> Self

Set commiment options.

pub fn signer(self, signer: &'a dyn Signer) -> Self

Add a signer to the signer list.

pub fn owned_signer(self, signer: impl Signer + Clone + 'static) -> Self

Add a owned sigenr to the signer list.

pub fn program(self, program_id: Pubkey) -> Self

Set program id.

pub fn accounts(self, accounts: Vec<AccountMeta>) -> Self

Append accounts for the main instruction.

pub fn anchor_accounts(self, accounts: impl ToAccountMetas) -> Self

Append accounts for the main instruction.

pub fn args(self, args: Vec<u8>) -> Self

Set arguments for the main instruction.

pub fn anchor_args(self, args: impl InstructionData) -> Self

Set arguments for the main instruction.

pub fn compute_budget(self, budget: ComputeBudget) -> Self

Set compute budget.

pub fn compute_budget_mut(&mut self) -> &mut ComputeBudget

Get mutable reference to the compute budget.

pub fn take_instruction(&mut self) -> Option<Instruction>

Take and construct the “main” instruction if present.

pub fn instructions(&self) -> Vec<Instruction>

Construct all instructions.

pub fn instructions_with_options( &self, without_compute_budget: bool, compute_unit_price_micro_lamports: Option<u64>, ) -> Vec<Instruction>

Construct all instructions with options.

pub fn get_output(&self) -> &T

Get the output.

pub fn swap_output<U>(self, output: U) -> (TransactionBuilder<'a, C, U>, T)

Set the output and return the previous.

pub fn output<U>(self, output: U) -> TransactionBuilder<'a, C, U>

Set the output.

pub fn clear_output(self) -> TransactionBuilder<'a, C, ()>

Clear the output.

pub fn pre_instruction(self, ix: Instruction) -> Self

Insert an instruction before the “main” instruction.

pub fn pre_instructions(self, ixs: Vec<Instruction>) -> Self

Insert instructions before the “main” instruction.

pub fn lookup_table(self, account: AddressLookupTableAccount) -> Self

Insert an address lookup table account.

pub fn lookup_tables( self, tables: impl IntoIterator<Item = (Pubkey, Vec<Pubkey>)>, ) -> Self

Insert many address lookup tables.

pub fn message_with_blockhash_and_options( &self, latest_hash: Hash, without_compute_budget: bool, compute_unit_price_micro_lamports: Option<u64>, ) -> Result<VersionedMessage>

Get versioned message with the given hash and options.

pub async fn message_with_options( &self, without_compute_budget: bool, compute_unit_price_micro_lamports: Option<u64>, ) -> Result<VersionedMessage>

Get versioned message with options.

pub fn signed_transaction_with_blockhash_and_options( &self, latest_hash: Hash, without_compute_budget: bool, compute_unit_price_micro_lamports: Option<u64>, ) -> Result<VersionedTransaction>

Get signed transaction with blockhash and options.

pub async fn signed_transaction_with_options( &self, without_compute_budget: bool, compute_unit_price_micro_lamports: Option<u64>, ) -> Result<VersionedTransaction>

Get signed transactoin with options.

pub async fn send_with_options( &self, without_compute_budget: bool, compute_unit_price_micro_lamports: Option<u64>, config: RpcSendTransactionConfig, ) -> Result<WithSlot<Signature>>

Sign and send the transaction with options.

pub async fn send_without_preflight(self) -> Result<Signature>

Build and send the transaction without preflight.

pub async fn send(self) -> Result<Signature>

Build and send the transaction with default options.

pub fn get_complete_lookup_table(&self) -> HashSet<Pubkey>

Get complete lookup table.

pub fn get_luts(&self) -> &HashMap<Pubkey, Vec<Pubkey>>

Get luts.

pub fn transaction_size(&self, is_versioned_transaction: bool) -> usize

Estimates the size of the result transaction.

See transaction_size() for more information.

pub async fn estimate_execution_fee( &self, _client: &RpcClient, compute_unit_price_micro_lamports: Option<u64>, ) -> Result<u64>

Estimated the execution fee of the result transaction.